Enemy's getDamageDealt is now float

+++ src/stray/entity/types/Enemy.java
@@ -8,6 +8,6 @@ package stray.entity.types;
  */
 public interface Enemy {
 
-	public abstract int getDamageDealt();
+	public abstract float getDamageDealt();
 	
 }
